server/requirements.txt
Flask==2.2.3
Flask-APScheduler==1.12.4
langchain==0.0.191
openai==0.27.8
faiss-cpu==1.7.4
python-dotenv==1.0.0
Flask-Cors==3.0.10
requests==2.31.0